+++ gradle.properties
@@ -1,4 +1,4 @@
 nablarchRepoReferenceUrl=https://oss.sonatype.org
-nablarchGradlePluginVersion=0.0.18
+nablarchGradlePluginVersion=0.0.20
 
 developLibUrl=https://oss.sonatype.org/content/groups/staging
