--- main/build.properties
@@ -1,38 +1,0 @@
-#common
-src.encoding=UTF-8
-filename.encoding=UTF-8
-
-#directory
-src.dir=java
-tool_lib.dir=tool_lib
-target.dir=target
-report.dir=reports
-resource.dir=resources
-lib.dir=lib
-
-#compile
-debug=on
-
-#jar
-jar.name=nablarch-wmq.jar
-zip.name=nablarch-wmq-src.zip
-
-#javadoc
-javadoc.dir=javadoc
-javadoc.window.title=Nablarch WebSphere MQ
-exclude.package.names=
-
-#checkstyle
-checkstyle.dir=${tool_lib.dir}/checkstyle-5.1
-checkstyle.jar=checkstyle-all-5.1.jar
-checkstyle.rule.file=${checkstyle.dir}/ssd_checks.xml
-checkstyle.report.dir=${report.dir}/checkstyle
-checkstyle.report.file=checkstyle-report.xml
-
-#findbugs
-findbugs.lib.dir=${tool_lib.dir}/findbugs-1.3.9-rc1
-findbugs.report.dir=${report.dir}/findbugs
-findbugs.report.file=findbugs-report.xml
-findbugs.published.config=tool_lib/published-config/production
-findbugs.exclude.filter=ssd_findbugs_exclude.xml
-
