--- main/build.xml
@@ -1,142 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<project name="nablarch_wmq" default="compile" basedir=".">
-
-    <property file="build.properties" />
-
-    <path id="lib.path">
-        <fileset dir="${lib.dir}">
-            <include name="**/*.jar" />
-        </fileset>
-    </path>
-
-    <path id="checkstyle.path" path="${checkstyle.dir}/${checkstyle.jar}" />
-
-    <target name="all"
-            depends="clean,compile,javadoc,checkstyle,findbugs,jar" />
-
-    <target name="release" depends="clean,compile,javadoc,jar,src.zip" />
-
-    <target name="clean">
-        <delete dir="${target.dir}" />
-        <delete dir="${javadoc.dir}" />
-        <delete dir="${report.dir}" />
-        <delete file="${jar.name}" />
-        <delete file="${zip.name}" />
-    </target>
-
-    <target name="compile">
-        <echo message="java compile version - ${java.version}" />
-        <delete dir="${target.dir}" />
-        <mkdir dir="${target.dir}" />
-        <javac destdir="${target.dir}"
-               target="1.5"
-               source="1.5"
-               srcdir="${src.dir}"
-               includes="**/*.java"
-               debug="${debug}"
-               failonerror="yes"
-               encoding="${src.encoding}"
-               includeantruntime="no">
-            <classpath>
-                <path refid="lib.path" />
-            </classpath>
-        </javac>
-    </target>
-
-    <target name="jar" depends="compile">
-        <delete file="${jar.name}" />
-        <jar destfile="${jar.name}">
-            <manifest>
-                <attribute name="Implementation-Version"
-                           value="${fw.version}" />
-            </manifest>
-            <fileset dir="${target.dir}" />
-            <fileset dir="${resource.dir}" />
-        </jar>
-    </target>
-
-    <target name="src.zip">
-        <delete file="${zip.name}" />
-        <zip destfile="${zip.name}" encoding="${filename.encoding}">
-            <fileset dir="${src.dir}" excludes="META-INF/" />
-            <fileset dir="${resource.dir}" />
-        </zip>
-    </target>
-
-    <target name="javadoc">
-        <delete dir="${javadoc.dir}" />
-        <mkdir dir="${javadoc.dir}" />
-        <javadoc destdir="${javadoc.dir}"
-                 encoding="${src.encoding}"
-                 author="true"
-                 version="true"
-                 access="protected"
-                 use="true"
-                 nonavbar="false"
-                 notree="true"
-                 nodeprecated="false"
-                 nodeprecatedlist="false"
-                 noindex="false"
-                 splitindex="true"
-                 nohelp="true"
-                 windowtitle="${javadoc.window.title}"
-                 docencoding="UTF-8"
-                 charset="UTF-8"
-                 defaultexcludes="yes"
-                 excludepackagenames="${exclude.package.names}"
-                 failOnError="yes">
-            <link offline="true"
-                  href="http://docs.oracle.com/javase/jp/6/api/"
-                  packagelistLoc="./" />
-            <doctitle>${javadoc.window.title}</doctitle>
-            <classpath>
-                <path refid="lib.path" />
-            </classpath>
-            <sourcepath path="${src.dir}" />
-            <package name="nablarch.integration.*" />
-            <group title="${javadoc.window.title}" packages="nablarch.integration.*" />
-        </javadoc>
-    </target>
-
-    <target name="checkstyle">
-        <taskdef resource="checkstyletask.properties"
-                 classpathref="checkstyle.path" />
-        <delete dir="${checkstyle.report.dir}" />
-        <mkdir dir="${checkstyle.report.dir}" />
-        <checkstyle config="${checkstyle.rule.file}"
-                    failureProperty="checkstyle.failure"
-                    failOnViolation="false">
-        	<property key="config_loc" value="${basedir}"/>
-            <fileset dir="${src.dir}" includes="**/*.java" />
-            <formatter type="xml"
-                       toFile="${checkstyle.report.dir}/${checkstyle.report.file}"
-                       usefile="true" />
-        	<classpath >
-        		<path path="${target.dir}"/>
-        		<path refid="lib.path"/>
-        	</classpath>
-        </checkstyle>
-    </target>
-
-    <target name="findbugs" depends="compile">
-        <taskdef name="findBugs"
-                 classname="edu.umd.cs.findbugs.anttask.FindBugsTask"
-                 classpath="${findbugs.lib.dir}/findbugs-ant.jar" />
-        <delete dir="${findbugs.report.dir}" />
-        <mkdir dir="${findbugs.report.dir}" />
-        <findBugs home="${findbugs.lib.dir}"
-                  output="xml"
-                  outputFile="${findbugs.report.dir}/${findbugs.report.file}"
-                  jvmargs="-Xms512M -Xmx512M"
-                  excludeFilter="${findbugs.exclude.filter}">
-            <class location="${target.dir}" />
-            <auxClasspath>
-                <fileset dir="${lib.dir}">
-                    <include name="**/*.jar" />
-                </fileset>
-            </auxClasspath>
-            <sourcePath path="${src.dir}" />
-            <systemProperty name="nablarch-findbugs-config" value="${findbugs.published.config}" />
-        </findBugs>
-    </target>
-</project>
