--- main/nablarch.jardesc
@@ -1,30 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" standalone="no"?>
-<jardesc>
-    <jar path="Nablarch_wmq/lib/nablarch.jar"/>
-    <options buildIfNeeded="true" compress="true" descriptionLocation="/Nablarch_wmq/nablarch.jardesc" exportErrors="true" exportWarnings="true" includeDirectoryEntries="false" overwrite="false" saveDescription="true" storeRefactorings="false" useSourceFolders="false"/>
-    <storedRefactorings deprecationInfo="true" structuralOnly="false"/>
-    <selectedProjects/>
-    <manifest generateManifest="true" manifestLocation="" manifestVersion="1.0" reuseManifest="false" saveManifest="false" usesManifest="true">
-        <sealing sealJar="false">
-            <packagesToSeal/>
-            <packagesToUnSeal/>
-        </sealing>
-    </manifest>
-    <selectedElements exportClassFiles="true" exportJavaFiles="false" exportOutputFolder="false">
-        <file path="/Nablarch/build.properties"/>
-        <file path="/Nablarch/publishapi.properties"/>
-        <file path="/Nablarch/.checkstyle"/>
-        <file path="/Nablarch/publishapi.xml"/>
-        <file path="/Nablarch/ssd_checks.xml"/>
-        <file path="/Nablarch/ssd_findbugs_exclude.xml"/>
-        <javaElement handleIdentifier="=Nablarch/java"/>
-        <file path="/Nablarch/.classpath"/>
-        <file path="/Nablarch/build.xml"/>
-        <file path="/Nablarch/ssd_checks_suppressions.xml"/>
-        <file path="/Nablarch/.project"/>
-        <file path="/Nablarch/update-schema.sh"/>
-        <file path="/Nablarch/Nablarch.iml"/>
-        <javaElement handleIdentifier="=Nablarch/resources"/>
-        <file path="/Nablarch/.fbprefs"/>
-    </selectedElements>
-</jardesc>
