--- main/ssd_findbugs_exclude.xml
@@ -1,10 +1,0 @@
-<FindBugsFilter>
-
-    <!-- インスタンスメソッドからstaticフィールドへの書き込み妥当性確認（済み） -->
-    <Match>
-      <Class name="nablarch.integration.messaging.wmq.provider.WmqMessagingProvider" />
-      <Method name="initialize" params="" returns="void" />
-      <Bug pattern="ST_WRITE_TO_STATIC_FROM_INSTANCE_METHOD" />
-    </Match>
-
-</FindBugsFilter>
