+++ src/main/java/nablarch/integration/messaging/wmq/package-info.java
