+++ src/main/java/nablarch/integration/messaging/wmq/xa/package-info.java
