+++ src/test/java/nablarch/integration/messaging/wmq/OnMemoryLogWriter.java
