+++ src/test/resources/nablarch/integration/messaging/wmq/provider/BasicWmqMqmdFieldsOperatorTest.xml
