--- test/build.properties
@@ -1,46 +1,0 @@
-#common
-src.encoding=UTF-8
-
-#directory
-test.src.dir=java
-test.resource.dir=resources
-test.target.dir=target
-test.lib.dir=lib
-test.tool.lib.dir=tool_lib
-
-main.root=../Nablarch_wmq
-main.resource.dir=${main.root}/resources
-main.src.dir=${main.root}/java
-main.lib.dir=${main.root}/lib
-main.target.dir=${main.root}/target
-
-copied.main.dir=copied_main
-copied.main.resource.dir=${copied.main.dir}/resources
-copied.main.src.dir=${copied.main.dir}/java
-copied.main.lib.dir=${copied.main.dir}/lib
-copied.main.target.dir=${copied.main.dir}/target
-
-unitTest.test.dir.include=**/*Test.class
-unitTest.test.dir.exclude=**/*TestBase.class
-
-report.dir=reports
-junit.report.dir=${report.dir}/junit
-cobertura.report.dir=${report.dir}/cobertura
-
-cobertura.dir=${test.tool.lib.dir}/cobertura-1.9.4.1
-instrumented.dir=instrumented
-
-#\u00e3\u0083\u0086\u00e3\u0082\u00b9\u00e3\u0083\u0088\u00e5\u00a4\u00b1\u00e6\u0095\u0097\u00e6\u0099\u0082\u00e3\u0081\u00ab\u00e3\u0083\u0093\u00e3\u0083\u00ab\u00e3\u0083\u0089\u00e3\u0083\u0097\u00e3\u0083\u00ad\u00e3\u0082\u00bb\u00e3\u0082\u00b9\u00e3\u0082\u0092\u00e6\u00ad\u00a2\u00e3\u0082\u0081\u00e3\u0082\u008b
-haltonfailure=off
-
-#tool_lib
-tool_lib.dir=tool_lib
-
-debug=true
-
-#findbugs
-findbugs.lib.dir=${tool_lib.dir}/findbugs-1.3.9-rc1
-findbugs.report.dir=${report.dir}/findbugs
-findbugs.report.file=findbugs-report.xml
-findbugs.include.filter=ssd_findbugs_include_for_test.xml
-findbugs.published.config=tool_lib/published-config/test
