--- test/resources/db.properties
@@ -1,3 +1,0 @@
-db.url =jdbc:oracle:thin:@localhost:1521/xe
-db.user = ssd
-db.password = ssd
