--- test/ssd_findbugs_include_for_test.xml
@@ -1,6 +1,0 @@
-<FindBugsFilter>
-    <!-- 使用不許可APIチェック -->
-    <Match>
-        <Bug pattern="UPU_UNPUBLISHED_API_USAGE"/>
-    </Match>
-</FindBugsFilter>
